functions/receiver/requirements.txt
functions-framework==3.*
slack-bolt>=1.18.0
google-cloud-tasks>=2.13.0
